Add Google Sheets import endpoint
