+++ package.json
@@ -25,6 +25,7 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.21.3",
     "react-scripts": "5.0.1",
+    "dotenv": "^16.4.5",
     "stylis": "^4.3.1",
     "stylis-plugin-rtl": "^2.1.1",
     "typescript": "^4.9.5",
